Move `PerfDataTable` to `Icinga\Module\Icingadb\Widget\Detail`
