+++ library/Icingadb/Widget/Detail/ObjectDetail.php
@@ -29,7 +29,7 @@ use Icinga\Module\Icingadb\Widget\ItemList\DowntimeList;
 use Icinga\Module\Icingadb\Widget\EmptyState;
 use ipl\Web\Widget\HorizontalKeyValue;
 use Icinga\Module\Icingadb\Widget\ItemList\CommentList;
-use Icinga\Module\Icingadb\Widget\PerfDataTable;
+use Icinga\Module\Icingadb\Widget\Detail\PerfDataTable;
 use Icinga\Module\Icingadb\Widget\PluginOutputContainer;
 use Icinga\Module\Icingadb\Widget\ShowMore;
 use Icinga\Module\Icingadb\Widget\TagList;
